Fix a bug in GetBaseName function
The implementation allocates memory for the base name result. And puts
a null char at the end of the string. But it doesn't take into account
that malloc can generate a memory slice containing null chars. When
this happens, the returned string is just the base name, without space
for the file extensions.

The fix just uses `memset()` to fill the string with blank spaces to
ensure no null characters are present other than the one inserted at
the end.