Add -w command line option to force Windows line endings
Otherwise line endings will follow the platform default. This is a breaking change, however generally we would expect the file layout to following the platform default (CRLF on Windows or LF on Unix); the need to force Windows line endings was for a special situation - compiling on Mac and using a programmer that expects Windows line endings. The JEDEC spec allows for CR and/or LF.